bionemo-recipes/recipes/mixtral_native_te/requirements.txt
datasets
hydra-core
torch
torchao!=0.14.0
torchdata
torchmetrics
tqdm
transformer_engine[pytorch]
transformers
wandb
zstandard
nvdlfw_inspect @ git+https://github.com/NVIDIA/nvidia-dlfw-inspect
